fix crash with data type

+++ app/src/main/java/com/example/spacex_app/presentation/model/dragonModel/DiameterModel.kt
@@ -6,5 +6,5 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class DiameterModel(
     var meters: Double? = 0.0,
-    var feet: Int? = 0
+    var feet: Double? = 0.0
 ) : Parcelable
